requirements.txt
requests
pandas
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
termcolor